Move restricted access logic to UserRestrictedAccessStore
