+++ src/__tests__/App.test.js
@@ -14,7 +14,6 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-
 import React from "react";
 import { render } from "@testing-library/react";
 
@@ -84,6 +83,7 @@ describe("App tests", () => {
     const user = { [metadataField]: { state_code: US_MO } };
     useRootStore.mockReturnValue({
       userStore: { user, isAuthorized: true },
+      userRestrictedAccessStore: { isLoading: false },
       currentTenantId: US_MO,
     });
 
@@ -99,6 +99,7 @@ describe("App tests", () => {
 
     useRootStore.mockReturnValue({
       userStore: { user, isAuthorized: true },
+      userRestrictedAccessStore: { isLoading: false },
       currentTenantId: US_ND,
     });
 
@@ -114,6 +115,7 @@ describe("App tests", () => {
     const user = { [metadataField]: { state_code: US_PA } };
     useRootStore.mockReturnValue({
       userStore: { user, isAuthorized: true },
+      userRestrictedAccessStore: { isLoading: false },
       currentTenantId: US_PA,
     });
 
@@ -129,6 +131,7 @@ describe("App tests", () => {
       window.history.pushState({}, "", "/some/page");
       useRootStore.mockReturnValue({
         userStore: { user: {}, isAuthorized: true },
+        userRestrictedAccessStore: { isLoading: false },
         currentTenantId: US_PA,
       });
 
@@ -142,6 +145,7 @@ describe("App tests", () => {
       window.history.pushState({}, "", "/some/page");
       useRootStore.mockReturnValue({
         userStore: { user: {}, isAuthorized: true },
+        userRestrictedAccessStore: { isLoading: false },
         currentTenantId: "US_XX",
       });
 
@@ -155,6 +159,7 @@ describe("App tests", () => {
   it("should render Loading component while user is loading", () => {
     useRootStore.mockReturnValue({
       userStore: { user: {}, userIsLoading: true, authorize: () => {} },
+      userRestrictedAccessStore: { isLoading: true },
     });
 
     const { container } = render(<App />);
@@ -165,7 +170,8 @@ describe("App tests", () => {
 
   it("should render the Error component if there is an error", () => {
     useRootStore.mockReturnValue({
-      userStore: { isLoading: false, isAuthorized: false },
+      userStore: { userIsLoading: false, isAuthorized: false },
+      userRestrictedAccessStore: { isLoading: false },
     });
 
     // do not log the expected error - keep tests less verbose
@@ -180,6 +186,7 @@ describe("App tests", () => {
     window.history.pushState({}, "", "/verify");
     useRootStore.mockReturnValue({
       userStore: { user: {}, isAuthorized: true },
+      userRestrictedAccessStore: { isLoading: false },
       currentTenantId: US_PA,
     });
